Add serialization support to Syzygy
As a part of efforts for issue 37, we will need the ability to easily serialize objects to and from binary streams in a lightweight manner. This CL adds a serialization engine to core_lib, with a simple native binary backend.

A first client of this functionality is the instrumenter, which uses serialization to write a small piece of metadata to an instrumented binary, linking it to the original binary it instrumented.

BUG=http://code.google.com/p/sawbuck/issues/detail?id=37
Review URL: http://codereview.appspot.com/4625072