Use section contributions for initial parsing of .text
This CL modifies Decomposer to use section contributions for the first pass of parsing, followed by the use of function and data symbols for refinement. The hope was that this would free us from control-flow analysis and block merging, but unfortunately this is still required in order to properly parse NaCl blocks (whose section contributions are too fine grained).

As a side-effect, section contributions give us reliable information regarding padding blocks in .text. This CL also introduces padding block detection, and modifies the two relinkers to not output padding blocks when reordering a section.
Review URL: http://codereview.appspot.com/4805044