Add JSON Stream Writer to core_lib.
This adds a utility for outputting formatted JSON directly to disk rather than via a Value and String intermediary. We are now outputting JSON in various locations (Decomposer, Metadata, SyzygyVersion, wsdump), so the intent is to migrate these to use the new functionality.

This also updates wsdump to use the new functionality.
Review URL: http://codereview.appspot.com/5161042