Create 'PrepareHeadersTransform'.
This transform is responsible for making sure that the DOS headers are trimmed and the NT headers are sufficiently large for the number of sections actually in the block-graph. To be run before layout, to ensure that the layout is valid and ready for FinalizeHeaders.

BUG=
TEST=

Review URL: http://codereview.appspot.com/5432069