Create WritablePdbStream.
This CL creates WritablePdbStream, and augments PdbStream with a 'factory' for getting access to the writable interface, if it is supported. This is not strictly an interface right now, but actually inherits from BufferWriter as a hack. Judicious comments explain the current situation, but eventually we'll want to hoist some proper interfaces here.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6164044