Hoist AddBasicBlockFrequencyDataTransform from coverage transform.
This CL creates a reusable transform to add the static basic-block
frequency to an image. A subsequent CL will transition the coverage
transform to reuse this new helper transform.

It also creates a reusable test fixture for transform unit-tests
that act on a decomposed test_dll block-graph. A subseqeunct CL
will update the other unit-tests to use the new fixture.

R = chrisha@chromium.org, siggi@chromium.org

Review URL: https://codereview.appspot.com/6490090