Small fix for BasicBlockLayoutTransform.
This is a fix for an issue that was only being encountered on the build bot with iterator debugging enabled. Basically, we were updating a sorted list in place and then doing queries to that list with lower_bound and upper_bound. With debugging enabled, it would blow up as the list slowly becomes unsorted as we update it in place. (We're sorting by a pointer which we then replace. This fix duplicates the pointer, and sorts/searches based on the copy. The original is updated in place.)

BUG=

Review URL: https://codereview.appspot.com/6655058