Create a simple logging service framwwork.
This CL creates a simple logger and infrastructure to hook it up to receive
messages over RPC. In this initial implementation it only logs text.

This is in prep for getting ASAN error logging out of chrome sandboxed
processes. In subsequent CLs we'll look at:

- wrapping this into a command-line app that sets up the logging
  environment then runs a specified executable under logging.

- adding stack trace support (with symbol resolution done in the logger).

R= siggi@chromium.org, sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/6816108