Small fix for bug in pdbfind_unittests.
It's possible for their to be multiple paths to the same file so checking for identical paths is not always sufficient. This centralizes EXPECT_SAME_FILE (which looks at the actual location of the file on disk) to core/unittest_util and reworks pdbfind_unittests to use it.

BUG=

Review URL: https://codereview.appspot.com/7058076