Create core::AddressFilter.
This is the data structure that will be used to implement instrumentation time filtering. It represents a memory space as a set of non-empty discontiguous intervals of 'marked' regions. The marked regions will be used to indicate which portions of the original binary's address space should be instrumented.

BUG=

Review URL: https://codereview.appspot.com/7437053