Directly report asan errors via breakpad if available.
This CL makes the asan error reporting breakpad aware. If it can locate
the expected breakpad reporting function in the exports of the currently
running EXE it will construct an exception and call the reporting function.

This has the positive side-effect of removing any requirement for the asan_rtl
symbols to be pushed to breakpad or the symsrc servers for reporting
purposes.

This has the negative side-effect of elevating the context away from the
asan_rtl which is tracking things like alloc/free points and shadow ram.

Maybe these can be stuffed into the exception record as params?

R= chrisha@chromium.org, siggi@chromium.org, sebmarchand@chromium.org

BUG=

Review URL: https://codereview.appspot.com/8097044