Add a callback mechanism to allow HeapProxy to report errors that originate from within it.
HeapProxy was previously only capable of generating 2 error types at one distinct location, with the address of the error being implicit. This allows HeapProxy to report arbitrary errors to its parent AsanRuntime. This functionality will be necessary to support checksum validation of blocks expiring from the quarantine, as the error is associated with a block that has nothing to do with the original block being freed.

BUG=
R=etienneb@chromium.org, sebmarchand@chromium.org

Review URL: https://codereview.appspot.com/70860043