Add a SectionOffsetAddress class.
Pdb files contain address represented as segment/offset
pairs. This CL introduces a class to encapsulate this
concept.

It has the same interface as AddressImpl, except for the
operator- that accepts another address of the same type.
The class is a lightweight wrapper for 2 integers, which
can be freely copied.

BUG=
R=chrisha@chromium.org

Review URL: https://codereview.appspot.com/114610045